requirements.txt
pyspark==3.5.4
duckdb==1.1.3
confluent-kafka
requests
dbt-duckdb
streamlit
pandas
tabulate
python-dotenv
pyarrow
